+++ src/dlt_filesystem/source/api.py
@@ -3,6 +3,7 @@ from dlt_filesystem.source.impl.remote import (
     AzureSource,
     GCSSource,
     HDFSSource,
+    R2Source,
     S3Source,
     SFTPSource,
 )
@@ -12,6 +13,7 @@ __all__ = [
     "GCSSource",
     "HDFSSource",
     "LocalFilesystemSource",
+    "R2Source",
     "S3Source",
     "SFTPSource",
 ]